requirements.txt
langchain
streamlit
faiss-cpu
openai
langchain-groq
python-dotenv
